% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{changes.getStartPageToken}
\alias{changes.getStartPageToken}
\title{Gets the starting pageToken for listing future changes.}
\usage{
changes.getStartPageToken(
  driveId = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  teamDriveId = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.changes.getStartPageToken"]]$scopes),
  return_response = F
)
}
\arguments{
\item{driveId}{The ID of the shared drive for which the starting pageToken for listing future changes from that shared drive is returned.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{teamDriveId}{Deprecated use driveId instead.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
