% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{changes.list}
\alias{changes.list}
\title{Lists the changes for a user or shared drive.}
\usage{
changes.list(
  pageToken,
  driveId = NULL,
  includeCorpusRemovals = NULL,
  includeItemsFromAllDrives = NULL,
  includePermissionsForView = NULL,
  includeRemoved = NULL,
  includeTeamDriveItems = NULL,
  pageSize = NULL,
  restrictToMyDrive = NULL,
  spaces = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  teamDriveId = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.changes.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{pageToken}{The token for continuing a previous list request on the next page. This should be set to the value of 'nextPageToken' from the previous response or to the response from the getStartPageToken method.}

\item{driveId}{The shared drive from which changes are returned. If specified the change IDs will be reflective of the shared drive; use the combined drive ID and change ID as an identifier.}

\item{includeCorpusRemovals}{Whether changes should include the file resource if the file is still accessible by the user at the time of the request, even when a file was removed from the list of changes and there will be no further change entries for this file.}

\item{includeItemsFromAllDrives}{Whether both My Drive and shared drive items should be included in results.}

\item{includePermissionsForView}{Specifies which additional view's permissions to include in the response. Only 'published' is supported.}

\item{includeRemoved}{Whether to include changes indicating that items have been removed from the list of changes, for example by deletion or loss of access.}

\item{includeTeamDriveItems}{Deprecated use includeItemsFromAllDrives instead.}

\item{pageSize}{The maximum number of changes to return per page.}

\item{restrictToMyDrive}{Whether to restrict the results to changes inside the My Drive hierarchy. This omits changes to files such as those in the Application Data folder or shared files which have not been added to My Drive.}

\item{spaces}{A comma-separated list of spaces to query within the user corpus. Supported values are 'drive', 'appDataFolder' and 'photos'.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{teamDriveId}{Deprecated use driveId instead.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
