% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{comments.create}
\alias{comments.create}
\title{Creates a new comment on a file.}
\usage{
comments.create(
  fileId,
  anchor = NULL,
  author = NULL,
  content = NULL,
  createdTime = NULL,
  deleted = NULL,
  htmlContent = NULL,
  id = NULL,
  kind = NULL,
  modifiedTime = NULL,
  quotedFileContent = NULL,
  replies = NULL,
  resolved = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.comments.create"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{anchor}{A region of the document represented as a JSON string. See anchor documentation for details on how to define and interpret anchor properties.}

\item{author}{The author of the comment. The author's email address and permission ID will not be populated.}

\item{content}{The plain text content of the comment. This field is used for setting the content, while htmlContent should be displayed.}

\item{createdTime}{The time at which the comment was created (RFC 3339 date-time).}

\item{deleted}{Whether the comment has been deleted. A deleted comment has no content.}

\item{htmlContent}{The content of the comment with HTML formatting.}

\item{id}{The ID of the comment.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#comment".}

\item{modifiedTime}{The last time the comment or any of its replies was modified (RFC 3339 date-time).}

\item{quotedFileContent}{The file content to which the comment refers, typically within the anchor region. For a text file, for example, this would be the text at the location of the comment.}

\item{replies}{The full list of replies to the comment in chronological order.}

\item{resolved}{Whether the comment has been resolved by one of its replies.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
