% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{comments.get}
\alias{comments.get}
\title{Gets a comment by ID.}
\usage{
comments.get(
  commentId,
  fileId,
  includeDeleted = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.comments.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{commentId}{The ID of the comment.}

\item{fileId}{The ID of the file.}

\item{includeDeleted}{Whether to return deleted comments. Deleted comments will not include their original content.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
