% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{comments.list}
\alias{comments.list}
\title{Lists a file's comments.}
\usage{
comments.list(
  fileId,
  includeDeleted = NULL,
  pageSize = NULL,
  pageToken = NULL,
  startModifiedTime = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.comments.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{includeDeleted}{Whether to include deleted comments. Deleted comments will not include their original content.}

\item{pageSize}{The maximum number of comments to return per page.}

\item{pageToken}{The token for continuing a previous list request on the next page. This should be set to the value of 'nextPageToken' from the previous response.}

\item{startModifiedTime}{The minimum value of 'modifiedTime' for the result comments (RFC 3339 date-time).}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
