% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{drives.delete}
\alias{drives.delete}
\title{Permanently deletes a shared drive for which the user is an organizer. The shared drive cannot contain any untrashed items.}
\usage{
drives.delete(
  driveId,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.drives.delete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{driveId}{The ID of the shared drive.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
