% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{drives.list}
\alias{drives.list}
\title{Lists the user's shared drives.}
\usage{
drives.list(
  pageSize = NULL,
  pageToken = NULL,
  q = NULL,
  useDomainAdminAccess = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.drives.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{pageSize}{Maximum number of shared drives to return.}

\item{pageToken}{Page token for shared drives.}

\item{q}{Query string for searching shared drives.}

\item{useDomainAdminAccess}{Issue the request as a domain administrator; if set to true, then all shared drives of the domain in which the requester is an administrator are returned.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
