% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{drives.update}
\alias{drives.update}
\title{Updates the metadate for a shared drive.}
\usage{
drives.update(
  driveId,
  useDomainAdminAccess = NULL,
  backgroundImageFile = NULL,
  backgroundImageLink = NULL,
  capabilities = NULL,
  colorRgb = NULL,
  createdTime = NULL,
  hidden = NULL,
  id = NULL,
  kind = NULL,
  name = NULL,
  restrictions = NULL,
  themeId = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.drives.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{driveId}{The ID of the shared drive.}

\item{useDomainAdminAccess}{Issue the request as a domain administrator; if set to true, then the requester will be granted access if they are an administrator of the domain to which the shared drive belongs.}

\item{backgroundImageFile}{An image file and cropping parameters from which a background image for this shared drive is set. This is a write only field; it can only be set on drive.drives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.}

\item{backgroundImageLink}{A short-lived link to this shared drive's background image.}

\item{capabilities}{Capabilities the current user has on this shared drive.}

\item{colorRgb}{The color of this shared drive as an RGB hex string. It can only be set on a drive.drives.update request that does not set themeId.}

\item{createdTime}{The time at which the shared drive was created (RFC 3339 date-time).}

\item{hidden}{Whether the shared drive is hidden from default view.}

\item{id}{The ID of this shared drive which is also the ID of the top level folder of this shared drive.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#drive".}

\item{name}{The name of this shared drive.}

\item{restrictions}{A set of restrictions that apply to this shared drive or items inside this shared drive.}

\item{themeId}{The ID of the theme from which the background image and color will be set. The set of possible driveThemes can be retrieved from a drive.about.get response. When not specified on a drive.drives.create request, a random theme is chosen from which the background image and color are set. This is a write-only field; it can only be set on requests that don't set colorRgb or backgroundImageFile.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
