% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.delete}
\alias{files.delete}
\title{Permanently deletes a file owned by the user without moving it to the trash. If the file belongs to a shared drive the user must be an organizer on the parent. If the target is a folder, all descendants owned by the user are also deleted.}
\usage{
files.delete(
  fileId,
  enforceSingleParent = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.files.delete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{enforceSingleParent}{Deprecated. If an item is not in a shared drive and its last parent is deleted but the item itself is not, the item will be placed under its owner's root.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
