% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.emptyTrash}
\alias{files.emptyTrash}
\title{Permanently deletes all of the user's trashed files.}
\usage{
files.emptyTrash(
  enforceSingleParent = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.files.emptyTrash"]]$scopes),
  return_response = F
)
}
\arguments{
\item{enforceSingleParent}{Deprecated. If an item is not in a shared drive and its last parent is deleted but the item itself is not, the item will be placed under its owner's root.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
