% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.export}
\alias{files.export}
\title{Exports a Google Doc to the requested MIME type and returns the exported content. Please note that the exported content is limited to 10MB.}
\usage{
files.export(
  fileId,
  mimeType,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.files.export"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{mimeType}{The MIME type of the format requested for this export.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
