% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.generateIds}
\alias{files.generateIds}
\title{Generates a set of file IDs which can be provided in create or copy requests.}
\usage{
files.generateIds(
  count = NULL,
  space = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.files.generateIds"]]$scopes),
  return_response = F
)
}
\arguments{
\item{count}{The number of IDs to return.}

\item{space}{The space in which the IDs can be used to create new files. Supported values are 'drive' and 'appDataFolder'.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
