% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.list}
\alias{files.list}
\title{Lists or searches files.}
\usage{
files.list(
  corpora = NULL,
  corpus = NULL,
  driveId = NULL,
  includeItemsFromAllDrives = NULL,
  includePermissionsForView = NULL,
  includeTeamDriveItems = NULL,
  orderBy = NULL,
  pageSize = NULL,
  pageToken = NULL,
  q = NULL,
  spaces = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  teamDriveId = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.files.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{corpora}{Groupings of files to which the query applies. Supported groupings are: 'user' (files created by, opened by, or shared directly with the user), 'drive' (files in the specified shared drive as indicated by the 'driveId'), 'domain' (files shared to the user's domain), and 'allDrives' (A combination of 'user' and 'drive' for all drives where the user is a member). When able, use 'user' or 'drive', instead of 'allDrives', for efficiency.}

\item{corpus}{The source of files to list. Deprecated: use 'corpora' instead.}

\item{driveId}{ID of the shared drive to search.}

\item{includeItemsFromAllDrives}{Whether both My Drive and shared drive items should be included in results.}

\item{includePermissionsForView}{Specifies which additional view's permissions to include in the response. Only 'published' is supported.}

\item{includeTeamDriveItems}{Deprecated use includeItemsFromAllDrives instead.}

\item{orderBy}{A comma-separated list of sort keys. Valid keys are 'createdTime', 'folder', 'modifiedByMeTime', 'modifiedTime', 'name', 'name_natural', 'quotaBytesUsed', 'recency', 'sharedWithMeTime', 'starred', and 'viewedByMeTime'. Each key sorts ascending by default, but may be reversed with the 'desc' modifier. Example usage: ?orderBy=folder,modifiedTime desc,name. Please note that there is a current limitation for users with approximately one million files in which the requested sort order is ignored.}

\item{pageSize}{The maximum number of files to return per page. Partial or empty result pages are possible even before the end of the files list has been reached.}

\item{pageToken}{The token for continuing a previous list request on the next page. This should be set to the value of 'nextPageToken' from the previous response.}

\item{q}{A query for filtering the file results. See the "Search for Files" guide for supported syntax.}

\item{spaces}{A comma-separated list of spaces to query within the corpus. Supported values are 'drive', 'appDataFolder' and 'photos'.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{teamDriveId}{Deprecated use driveId instead.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
