% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.update}
\alias{files.update}
\title{Updates a file's metadata and/or content. This method supports patch semantics.}
\usage{
files.update(
  fileId,
  addParents = NULL,
  enforceSingleParent = NULL,
  includePermissionsForView = NULL,
  keepRevisionForever = NULL,
  ocrLanguage = NULL,
  removeParents = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  useContentAsIndexableText = NULL,
  appProperties = NULL,
  capabilities = NULL,
  contentHints = NULL,
  contentRestrictions = NULL,
  copyRequiresWriterPermission = NULL,
  createdTime = NULL,
  description = NULL,
  driveId = NULL,
  explicitlyTrashed = NULL,
  exportLinks = NULL,
  fileExtension = NULL,
  folderColorRgb = NULL,
  fullFileExtension = NULL,
  hasAugmentedPermissions = NULL,
  hasThumbnail = NULL,
  headRevisionId = NULL,
  iconLink = NULL,
  id = NULL,
  imageMediaMetadata = NULL,
  isAppAuthorized = NULL,
  kind = NULL,
  lastModifyingUser = NULL,
  md5Checksum = NULL,
  mimeType = NULL,
  modifiedByMe = NULL,
  modifiedByMeTime = NULL,
  modifiedTime = NULL,
  name = NULL,
  originalFilename = NULL,
  ownedByMe = NULL,
  owners = NULL,
  parents = NULL,
  permissionIds = NULL,
  permissions = NULL,
  properties = NULL,
  quotaBytesUsed = NULL,
  shared = NULL,
  sharedWithMeTime = NULL,
  sharingUser = NULL,
  shortcutDetails = NULL,
  size = NULL,
  spaces = NULL,
  starred = NULL,
  teamDriveId = NULL,
  thumbnailLink = NULL,
  thumbnailVersion = NULL,
  trashed = NULL,
  trashedTime = NULL,
  trashingUser = NULL,
  version = NULL,
  videoMediaMetadata = NULL,
  viewedByMe = NULL,
  viewedByMeTime = NULL,
  viewersCanCopyContent = NULL,
  webContentLink = NULL,
  webViewLink = NULL,
  writersCanShare = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.files.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{addParents}{A comma-separated list of parent IDs to add.}

\item{enforceSingleParent}{Deprecated. Adding files to multiple folders is no longer supported. Use shortcuts instead.}

\item{includePermissionsForView}{Specifies which additional view's permissions to include in the response. Only 'published' is supported.}

\item{keepRevisionForever}{Whether to set the 'keepForever' field in the new head revision. This is only applicable to files with binary content in Google Drive. Only 200 revisions for the file can be kept forever. If the limit is reached, try deleting pinned revisions.}

\item{ocrLanguage}{A language hint for OCR processing during image import (ISO 639-1 code).}

\item{removeParents}{A comma-separated list of parent IDs to remove.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{useContentAsIndexableText}{Whether to use the uploaded content as indexable text.}

\item{appProperties}{A collection of arbitrary key-value pairs which are private to the requesting app.Entries with null values are cleared in update and copy requests. These properties can only be retrieved using an authenticated request. An authenticated request uses an access token obtained with a OAuth 2 client ID. You cannot use an API key to retrieve private properties.}

\item{capabilities}{Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take.}

\item{contentHints}{Additional information about the content of the file. These fields are never populated in responses.}

\item{contentRestrictions}{Restrictions for accessing the content of the file. Only populated if such a restriction exists.}

\item{copyRequiresWriterPermission}{Whether the options to copy, print, or download this file, should be disabled for readers and commenters.}

\item{createdTime}{The time at which the file was created (RFC 3339 date-time).}

\item{description}{A short description of the file.}

\item{driveId}{ID of the shared drive the file resides in. Only populated for items in shared drives.}

\item{explicitlyTrashed}{Whether the file has been explicitly trashed, as opposed to recursively trashed from a parent folder.}

\item{exportLinks}{Links for exporting Docs Editors files to specific formats.}

\item{fileExtension}{The final component of fullFileExtension. This is only available for files with binary content in Google Drive.}

\item{folderColorRgb}{The color for a folder as an RGB hex string. The supported colors are published in the folderColorPalette field of the About resource.If an unsupported color is specified, the closest color in the palette will be used instead.}

\item{fullFileExtension}{The full file extension extracted from the name field. May contain multiple concatenated extensions, such as "tar.gz". This is only available for files with binary content in Google Drive.This is automatically updated when the name field changes, however it is not cleared if the new name does not contain a valid extension.}

\item{hasAugmentedPermissions}{Whether there are permissions directly on this file. This field is only populated for items in shared drives.}

\item{hasThumbnail}{Whether this file has a thumbnail. This does not indicate whether the requesting app has access to the thumbnail. To check access, look for the presence of the thumbnailLink field.}

\item{headRevisionId}{The ID of the file's head revision. This is currently only available for files with binary content in Google Drive.}

\item{iconLink}{A static, unauthenticated link to the file's icon.}

\item{id}{The ID of the file.}

\item{imageMediaMetadata}{Additional metadata about image media, if available.}

\item{isAppAuthorized}{Whether the file was created or opened by the requesting app.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#file".}

\item{lastModifyingUser}{The last user to modify the file.}

\item{md5Checksum}{The MD5 checksum for the content of the file. This is only applicable to files with binary content in Google Drive.}

\item{mimeType}{The MIME type of the file.Google Drive will attempt to automatically detect an appropriate value from uploaded content if no value is provided. The value cannot be changed unless a new revision is uploaded.If a file is created with a Google Doc MIME type, the uploaded content will be imported if possible. The supported import formats are published in the About resource.}

\item{modifiedByMe}{Whether the file has been modified by this user.}

\item{modifiedByMeTime}{The last time the file was modified by the user (RFC 3339 date-time).}

\item{modifiedTime}{The last time the file was modified by anyone (RFC 3339 date-time).Note that setting modifiedTime will also update modifiedByMeTime for the user.}

\item{name}{The name of the file. This is not necessarily unique within a folder. Note that for immutable items such as the top level folders of shared drives, My Drive root folder, and Application Data folder the name is constant.}

\item{originalFilename}{The original filename of the uploaded content if available, or else the original value of the name field. This is only available for files with binary content in Google Drive.}

\item{ownedByMe}{Whether the user owns the file. Not populated for items in shared drives.}

\item{owners}{The owners of the file. Currently, only certain legacy files may have more than one owner. Not populated for items in shared drives.}

\item{parents}{The IDs of the parent folders which contain the file.If not specified as part of a create request, the file will be placed directly in the user's My Drive folder. If not specified as part of a copy request, the file will inherit any discoverable parents of the source file. Update requests must use the addParents and removeParents parameters to modify the parents list.}

\item{permissionIds}{List of permission IDs for users with access to this file.}

\item{permissions}{The full list of permissions for the file. This is only available if the requesting user can share the file. Not populated for items in shared drives.}

\item{properties}{A collection of arbitrary key-value pairs which are visible to all apps.Entries with null values are cleared in update and copy requests.}

\item{quotaBytesUsed}{The number of storage quota bytes used by the file. This includes the head revision as well as previous revisions with keepForever enabled.}

\item{shared}{Whether the file has been shared. Not populated for items in shared drives.}

\item{sharedWithMeTime}{The time at which the file was shared with the user, if applicable (RFC 3339 date-time).}

\item{sharingUser}{The user who shared the file with the requesting user, if applicable.}

\item{shortcutDetails}{Shortcut file details. Only populated for shortcut files, which have the mimeType field set to application/vnd.google-apps.shortcut.}

\item{size}{The size of the file's content in bytes. This is applicable to binary files in Google Drive and Google Docs files.}

\item{spaces}{The list of spaces which contain the file. The currently supported values are 'drive', 'appDataFolder' and 'photos'.}

\item{starred}{Whether the user has starred the file.}

\item{teamDriveId}{Deprecated - use driveId instead.}

\item{thumbnailLink}{A short-lived link to the file's thumbnail, if available. Typically lasts on the order of hours. Only populated when the requesting app can access the file's content. If the file isn't shared publicly, the URL returned in Files.thumbnailLink must be fetched using a credentialed request.}

\item{thumbnailVersion}{The thumbnail version for use in thumbnail cache invalidation.}

\item{trashed}{Whether the file has been trashed, either explicitly or from a trashed parent folder. Only the owner may trash a file. The trashed item is excluded from all files.list responses returned for any user who does not own the file. However, all users with access to the file can see the trashed item metadata in an API response. All users with access can copy, download, export, and share the file.}

\item{trashedTime}{The time that the item was trashed (RFC 3339 date-time). Only populated for items in shared drives.}

\item{trashingUser}{If the file has been explicitly trashed, the user who trashed it. Only populated for items in shared drives.}

\item{version}{A monotonically increasing version number for the file. This reflects every change made to the file on the server, even those not visible to the user.}

\item{videoMediaMetadata}{Additional metadata about video media. This may not be available immediately upon upload.}

\item{viewedByMe}{Whether the file has been viewed by this user.}

\item{viewedByMeTime}{The last time the file was viewed by the user (RFC 3339 date-time).}

\item{viewersCanCopyContent}{Deprecated - use copyRequiresWriterPermission instead.}

\item{webContentLink}{A link for downloading the content of the file in a browser. This is only available for files with binary content in Google Drive.}

\item{webViewLink}{A link for opening the file in a relevant Google editor or viewer in a browser.}

\item{writersCanShare}{Whether users with only writer permission can modify the file's permissions. Not populated for items in shared drives.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.scripts
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
