% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files.watch}
\alias{files.watch}
\title{Subscribes to changes to a file}
\usage{
files.watch(
  fileId,
  acknowledgeAbuse = NULL,
  includePermissionsForView = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  address = NULL,
  expiration = NULL,
  id = NULL,
  kind = NULL,
  params = NULL,
  payload = NULL,
  resourceId = NULL,
  resourceUri = NULL,
  token = NULL,
  type = NULL,
  gargle_token = gargle::token_fetch(scopes = .endpoints[["drive.files.watch"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{acknowledgeAbuse}{Whether the user is acknowledging the risk of downloading known malware or other abusive files. This is only applicable when alt=media.}

\item{includePermissionsForView}{Specifies which additional view's permissions to include in the response. Only 'published' is supported.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{address}{The address where notifications are delivered for this channel.}

\item{expiration}{Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds. Optional.}

\item{id}{A UUID or similar unique string that identifies this channel.}

\item{kind}{Identifies this as a notification channel used to watch for changes to a resource, which is "api#channel".}

\item{params}{Additional parameters controlling delivery channel behavior. Optional.}

\item{payload}{A Boolean value to indicate whether payload is wanted. Optional.}

\item{resourceId}{An opaque ID that identifies the resource being watched on this channel. Stable across different API versions.}

\item{resourceUri}{A version-specific identifier for the watched resource.}

\item{token}{An arbitrary string delivered to the target address with each notification delivered over this channel. Optional.}

\item{type}{The type of delivery mechanism used for this channel.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
