% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions.delete}
\alias{permissions.delete}
\title{Deletes a permission.}
\usage{
permissions.delete(
  fileId,
  permissionId,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  useDomainAdminAccess = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.permissions.delete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file or shared drive.}

\item{permissionId}{The ID of the permission.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{useDomainAdminAccess}{Issue the request as a domain administrator; if set to true, then the requester will be granted access if the file ID parameter refers to a shared drive and the requester is an administrator of the domain to which the shared drive belongs.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
