% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{permissions.update}
\alias{permissions.update}
\title{Updates a permission with patch semantics.}
\usage{
permissions.update(
  fileId,
  permissionId,
  removeExpiration = NULL,
  supportsAllDrives = NULL,
  supportsTeamDrives = NULL,
  transferOwnership = NULL,
  useDomainAdminAccess = NULL,
  allowFileDiscovery = NULL,
  deleted = NULL,
  displayName = NULL,
  domain = NULL,
  emailAddress = NULL,
  expirationTime = NULL,
  id = NULL,
  kind = NULL,
  permissionDetails = NULL,
  photoLink = NULL,
  role = NULL,
  teamDrivePermissionDetails = NULL,
  type = NULL,
  view = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.permissions.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file or shared drive.}

\item{permissionId}{The ID of the permission.}

\item{removeExpiration}{Whether to remove the expiration date.}

\item{supportsAllDrives}{Whether the requesting application supports both My Drives and shared drives.}

\item{supportsTeamDrives}{Deprecated use supportsAllDrives instead.}

\item{transferOwnership}{Whether to transfer ownership to the specified user and downgrade the current owner to a writer. This parameter is required as an acknowledgement of the side effect.}

\item{useDomainAdminAccess}{Issue the request as a domain administrator; if set to true, then the requester will be granted access if the file ID parameter refers to a shared drive and the requester is an administrator of the domain to which the shared drive belongs.}

\item{allowFileDiscovery}{Whether the permission allows the file to be discovered through search. This is only applicable for permissions of type domain or anyone.}

\item{deleted}{Whether the account associated with this permission has been deleted. This field only pertains to user and group permissions.}

\item{displayName}{The "pretty" name of the value of the permission. The following is a list of examples for each type of permission:  - user - User's full name, as defined for their Google account, such as "Joe Smith." - group - Name of the Google Group, such as "The Company Administrators." - domain - String domain name, such as "thecompany.com." - anyone - No displayName is present.}

\item{domain}{The domain to which this permission refers.}

\item{emailAddress}{The email address of the user or group to which this permission refers.}

\item{expirationTime}{The time at which this permission will expire (RFC 3339 date-time). Expiration times have the following restrictions:  - They can only be set on user and group permissions - The time must be in the future - The time cannot be more than a year in the future}

\item{id}{The ID of this permission. This is a unique identifier for the grantee, and is published in User resources as permissionId. IDs should be treated as opaque values.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#permission".}

\item{permissionDetails}{Details of whether the permissions on this shared drive item are inherited or directly on this item. This is an output-only field which is present only for shared drive items.}

\item{photoLink}{A link to the user's profile photo, if available.}

\item{role}{The role granted by this permission. While new values may be supported in the future, the following are currently allowed:  - owner - organizer - fileOrganizer - writer - commenter - reader}

\item{teamDrivePermissionDetails}{Deprecated - use permissionDetails instead.}

\item{type}{The type of the grantee. Valid values are:  - user - group - domain - anyone  When creating a permission, if type is user or group, you must provide an emailAddress for the user or group. When type is domain, you must provide a domain. There isn't extra information required for a anyone type.}

\item{view}{Indicates the view for this permission. Only populated for permissions that belong to a view. published is the only supported value.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
