% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replies.R
\name{replies.create}
\alias{replies.create}
\title{Creates a new reply to a comment.}
\usage{
replies.create(
  commentId,
  fileId,
  action = NULL,
  author = NULL,
  content = NULL,
  createdTime = NULL,
  deleted = NULL,
  htmlContent = NULL,
  id = NULL,
  kind = NULL,
  modifiedTime = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.replies.create"]]$scopes),
  return_response = F
)
}
\arguments{
\item{commentId}{The ID of the comment.}

\item{fileId}{The ID of the file.}

\item{action}{The action the reply performed to the parent comment. Valid values are:  - resolve - reopen}

\item{author}{The author of the reply. The author's email address and permission ID will not be populated.}

\item{content}{The plain text content of the reply. This field is used for setting the content, while htmlContent should be displayed. This is required on creates if no action is specified.}

\item{createdTime}{The time at which the reply was created (RFC 3339 date-time).}

\item{deleted}{Whether the reply has been deleted. A deleted reply has no content.}

\item{htmlContent}{The content of the reply with HTML formatting.}

\item{id}{The ID of the reply.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#reply".}

\item{modifiedTime}{The last time the reply was modified (RFC 3339 date-time).}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
