% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{revisions.get}
\alias{revisions.get}
\title{Gets a revision's metadata or content by ID.}
\usage{
revisions.get(
  fileId,
  revisionId,
  acknowledgeAbuse = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.revisions.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{revisionId}{The ID of the revision.}

\item{acknowledgeAbuse}{Whether the user is acknowledging the risk of downloading known malware or other abusive files. This is only applicable when alt=media.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
