% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{revisions.list}
\alias{revisions.list}
\title{Lists a file's revisions.}
\usage{
revisions.list(
  fileId,
  pageSize = NULL,
  pageToken = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.revisions.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{pageSize}{The maximum number of revisions to return per page.}

\item{pageToken}{The token for continuing a previous list request on the next page. This should be set to the value of 'nextPageToken' from the previous response.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
\item https://www.googleapis.com/auth/drive.metadata
\item https://www.googleapis.com/auth/drive.metadata.readonly
\item https://www.googleapis.com/auth/drive.photos.readonly
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
