% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisions.R
\name{revisions.update}
\alias{revisions.update}
\title{Updates a revision with patch semantics.}
\usage{
revisions.update(
  fileId,
  revisionId,
  exportLinks = NULL,
  id = NULL,
  keepForever = NULL,
  kind = NULL,
  lastModifyingUser = NULL,
  md5Checksum = NULL,
  mimeType = NULL,
  modifiedTime = NULL,
  originalFilename = NULL,
  publishAuto = NULL,
  published = NULL,
  publishedLink = NULL,
  publishedOutsideDomain = NULL,
  size = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.revisions.update"]]$scopes),
  return_response = F
)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{revisionId}{The ID of the revision.}

\item{exportLinks}{Links for exporting Docs Editors files to specific formats.}

\item{id}{The ID of the revision.}

\item{keepForever}{Whether to keep this revision forever, even if it is no longer the head revision. If not set, the revision will be automatically purged 30 days after newer content is uploaded. This can be set on a maximum of 200 revisions for a file.This field is only applicable to files with binary content in Drive.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#revision".}

\item{lastModifyingUser}{The last user to modify this revision.}

\item{md5Checksum}{The MD5 checksum of the revision's content. This is only applicable to files with binary content in Drive.}

\item{mimeType}{The MIME type of the revision.}

\item{modifiedTime}{The last time the revision was modified (RFC 3339 date-time).}

\item{originalFilename}{The original filename used to create this revision. This is only applicable to files with binary content in Drive.}

\item{publishAuto}{Whether subsequent revisions will be automatically republished. This is only applicable to Docs Editors files.}

\item{published}{Whether this revision is published. This is only applicable to Docs Editors files.}

\item{publishedLink}{A link to the published revision. This is only populated for Google Sites files.}

\item{publishedOutsideDomain}{Whether this revision is published outside the domain. This is only applicable to Docs Editors files.}

\item{size}{The size of the revision's content in bytes. This is only applicable to files with binary content in Drive.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/drive.file
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
