% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamdrives.R
\name{teamdrives.create}
\alias{teamdrives.create}
\title{Deprecated use drives.create instead.}
\usage{
teamdrives.create(
  requestId,
  backgroundImageFile = NULL,
  backgroundImageLink = NULL,
  capabilities = NULL,
  colorRgb = NULL,
  createdTime = NULL,
  id = NULL,
  kind = NULL,
  name = NULL,
  restrictions = NULL,
  themeId = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.teamdrives.create"]]$scopes),
  return_response = F
)
}
\arguments{
\item{requestId}{An ID, such as a random UUID, which uniquely identifies this user's request for idempotent creation of a Team Drive. A repeated request by the same user and with the same request ID will avoid creating duplicates by attempting to create the same Team Drive. If the Team Drive already exists a 409 error will be returned.}

\item{backgroundImageFile}{An image file and cropping parameters from which a background image for this Team Drive is set. This is a write only field; it can only be set on drive.teamdrives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.}

\item{backgroundImageLink}{A short-lived link to this Team Drive's background image.}

\item{capabilities}{Capabilities the current user has on this Team Drive.}

\item{colorRgb}{The color of this Team Drive as an RGB hex string. It can only be set on a drive.teamdrives.update request that does not set themeId.}

\item{createdTime}{The time at which the Team Drive was created (RFC 3339 date-time).}

\item{id}{The ID of this Team Drive which is also the ID of the top level folder of this Team Drive.}

\item{kind}{Identifies what kind of resource this is. Value: the fixed string "drive#teamDrive".}

\item{name}{The name of this Team Drive.}

\item{restrictions}{A set of restrictions that apply to this Team Drive or items inside this Team Drive.}

\item{themeId}{The ID of the theme from which the background image and color will be set. The set of possible teamDriveThemes can be retrieved from a drive.about.get response. When not specified on a drive.teamdrives.create request, a random theme is chosen from which the background image and color are set. This is a write-only field; it can only be set on requests that don't set colorRgb or backgroundImageFile.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
