% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamdrives.R
\name{teamdrives.delete}
\alias{teamdrives.delete}
\title{Deprecated use drives.delete instead.}
\usage{
teamdrives.delete(
  teamDriveId,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.teamdrives.delete"]]$scopes),
  return_response = F
)
}
\arguments{
\item{teamDriveId}{The ID of the Team Drive}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
