% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamdrives.R
\name{teamdrives.get}
\alias{teamdrives.get}
\title{Deprecated use drives.get instead.}
\usage{
teamdrives.get(
  teamDriveId,
  useDomainAdminAccess = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.teamdrives.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{teamDriveId}{The ID of the Team Drive}

\item{useDomainAdminAccess}{Issue the request as a domain administrator; if set to true, then the requester will be granted access if they are an administrator of the domain to which the Team Drive belongs.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
