% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamdrives.R
\name{teamdrives.list}
\alias{teamdrives.list}
\title{Deprecated use drives.list instead.}
\usage{
teamdrives.list(
  pageSize = NULL,
  pageToken = NULL,
  q = NULL,
  useDomainAdminAccess = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["drive.teamdrives.list"]]$scopes),
  return_response = F
)
}
\arguments{
\item{pageSize}{Maximum number of Team Drives to return.}

\item{pageToken}{Page token for Team Drives.}

\item{q}{Query string for searching Team Drives.}

\item{useDomainAdminAccess}{Issue the request as a domain administrator; if set to true, then all Team Drives of the domain in which the requester is an administrator are returned.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{make_google_package}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
}
}
\seealso{
\href{https://developers.google.com/drive/}{Google Documentation}
}
